--- a/workers/api-gateway-r2.ts
+++ b/workers/api-gateway-r2.ts
@@ -1,0 +1,167 @@
+/**
+ * Cloudflare Worker - API Gateway with R2 Integration
+ * R2から直接ランキングデータを配信
+ */
+
+import { gzipSync, gunzipSync } from 'zlib'
+import type { RankingData } from '../types/ranking'
+
+interface Env {
+  VERCEL_DEPLOYMENT_URL: string
+  WORKER_AUTH_KEY: string
+  R2_BUCKET: R2Bucket
+}
+
+// セキュリティヘッダー定義
+const securityHeaders = {
+  'Content-Security-Policy': "default-src 'self'; script-src 'self' 'unsafe-inline' https://*.vercel-scripts.com; style-src 'self' 'unsafe-inline'; img-src 'self' https: data: blob:; font-src 'self' data:; connect-src 'self' https:; frame-ancestors 'none'; base-uri 'self'; form-action 'self'; media-src 'self' https:; object-src 'none'",
+  'Strict-Transport-Security': 'max-age=31536000; includeSubDomains',
+  'X-Frame-Options': 'DENY',
+  'X-Content-Type-Options': 'nosniff',
+  'X-XSS-Protection': '1; mode=block',
+  'Referrer-Policy': 'strict-origin-when-cross-origin',
+  'Permissions-Policy': 'accelerometer=(), camera=(), geolocation=(), gyroscope=(), magnetometer=(), microphone=(), payment=(), usb=()',
+  'Cross-Origin-Opener-Policy': 'same-origin',
+  'X-DNS-Prefetch-Control': 'on'
+}
+
+// CORSヘッダー定義
+const corsHeaders = {
+  'Access-Control-Allow-Origin': '*',
+  'Access-Control-Allow-Methods': 'GET, OPTIONS',
+  'Access-Control-Allow-Headers': 'Content-Type',
+  'Access-Control-Max-Age': '86400'
+}
+
+export default {
+  async fetch(request: Request, env: Env, ctx: ExecutionContext): Promise<Response> {
+    const url = new URL(request.url)
+    
+    // OPTIONS リクエストの処理
+    if (request.method === 'OPTIONS') {
+      return new Response(null, {
+        status: 204,
+        headers: corsHeaders
+      })
+    }
+    
+    // /api/ranking パスの処理（R2から直接配信）
+    if (url.pathname === '/api/ranking' && env.R2_BUCKET) {
+      try {
+        // クエリパラメータからジャンルと期間を取得
+        const genre = url.searchParams.get('genre') || 'all'
+        const period = url.searchParams.get('period') || '24h'
+        
+        // キャッシュキー
+        const cacheKey = new Request(`https://r2-cache.nico-rank.com/ranking/${genre}/${period}`, request)
+        const cache = caches.default
+        
+        // キャッシュチェック
+        let response = await cache.match(cacheKey)
+        if (response) {
+          response = new Response(response.body, response)
+          response.headers.set('X-Cache-Status', 'HIT')
+          return response
+        }
+        
+        // R2から読み取り
+        const r2Key = `rankings/${genre}/${period}.json`
+        const r2Object = await env.R2_BUCKET.get(r2Key)
+        
+        if (!r2Object) {
+          // R2にデータがない場合はVercelにフォールバック
+          console.log(`R2 miss for ${r2Key}, falling back to Vercel`)
+          return proxyToVercel(request, env)
+        }
+        
+        // R2から取得したデータを返す
+        const data = await r2Object.text()
+        response = new Response(data, {
+          status: 200,
+          headers: {
+            'Content-Type': 'application/json',
+            'Cache-Control': 'public, max-age=1800, s-maxage=3600',
+            'X-Data-Source': 'r2-direct',
+            'X-Cache-Status': 'MISS',
+            ...corsHeaders,
+            ...securityHeaders
+          }
+        })
+        
+        // キャッシュに保存
+        ctx.waitUntil(cache.put(cacheKey, response.clone()))
+        
+        return response
+        
+      } catch (error) {
+        console.error('R2 read error:', error)
+        // エラー時はVercelにフォールバック
+        return proxyToVercel(request, env)
+      }
+    }
+    
+    // その他のリクエストはVercelへプロキシ
+    return proxyToVercel(request, env)
+  }
+}
+
+// Vercelへのプロキシ関数
+async function proxyToVercel(request: Request, env: Env): Promise<Response> {
+  const url = new URL(request.url)
+  const targetHost = env.VERCEL_DEPLOYMENT_URL || 'nico-ranking-custom-yjsns-projects.vercel.app'
+  
+  // プロキシ用のURL構築
+  const proxyUrl = new URL(url.pathname + url.search, `https://${targetHost}`)
+  
+  // リクエストヘッダーの準備
+  const headers = new Headers(request.headers)
+  headers.set('Host', targetHost)
+  headers.set('X-Forwarded-Host', url.hostname)
+  headers.set('X-Forwarded-Proto', 'https')
+  headers.set('X-Real-IP', request.headers.get('CF-Connecting-IP') || '')
+  
+  // 認証キーを追加
+  if (env.WORKER_AUTH_KEY) {
+    headers.set('X-Worker-Auth', env.WORKER_AUTH_KEY)
+  }
+  
+  // Vercelへのリクエスト
+  const proxyRequest = new Request(proxyUrl.toString(), {
+    method: request.method,
+    headers,
+    body: request.body,
+    redirect: 'manual'
+  })
+  
+  try {
+    const response = await fetch(proxyRequest)
+    
+    // レスポンスヘッダーの処理
+    const responseHeaders = new Headers(response.headers)
+    
+    // セキュリティヘッダーを追加
+    Object.entries(securityHeaders).forEach(([key, value]) => {
+      responseHeaders.set(key, value)
+    })
+    
+    // CORSヘッダーを追加
+    Object.entries(corsHeaders).forEach(([key, value]) => {
+      responseHeaders.set(key, value)
+    })
+    
+    return new Response(response.body, {
+      status: response.status,
+      statusText: response.statusText,
+      headers: responseHeaders
+    })
+  } catch (error) {
+    console.error('Proxy error:', error)
+    return new Response('Gateway Error', { 
+      status: 502,
+      headers: {
+        'Content-Type': 'text/plain',
+        ...corsHeaders
+      }
+    })
+  }
+}