--- a/workers/shared/security-headers.ts
+++ b/workers/shared/security-headers.ts
@@ -1,0 +1,40 @@
+/**
+ * 共通セキュリティヘッダー定義
+ */
+
+export const SECURITY_HEADERS = {
+  'Content-Security-Policy': "default-src 'self'; script-src 'self' 'unsafe-inline' https://*.vercel-scripts.com https://vercel.live; style-src 'self' 'unsafe-inline'; img-src 'self' data: https: blob:; font-src 'self' data:; connect-src 'self' https://*.niconico.jp https://*.nicovideo.jp; media-src 'self' https://*.niconico.jp https://*.nicovideo.jp; object-src 'none'; base-uri 'self'; form-action 'self'; frame-ancestors 'none'; upgrade-insecure-requests",
+  'Strict-Transport-Security': 'max-age=31536000; includeSubDomains',
+  'X-Frame-Options': 'DENY',
+  'X-Content-Type-Options': 'nosniff',
+  'X-XSS-Protection': '1; mode=block',
+  'Referrer-Policy': 'strict-origin-when-cross-origin',
+  'Permissions-Policy': 'camera=(), microphone=(), geolocation=()',
+  'Cross-Origin-Opener-Policy': 'same-origin',
+  'X-DNS-Prefetch-Control': 'on'
+}
+
+export const CORS_HEADERS = {
+  'Access-Control-Allow-Origin': '*',
+  'Access-Control-Allow-Methods': 'GET, OPTIONS',
+  'Access-Control-Allow-Headers': 'Content-Type',
+  'Access-Control-Max-Age': '86400'
+}
+
+/**
+ * レスポンスにセキュリティヘッダーを適用
+ */
+export function applySecurityHeaders(response: Response, url: URL): Response {
+  const newHeaders = new Headers(response.headers)
+  
+  // セキュリティヘッダーを適用
+  Object.entries(SECURITY_HEADERS).forEach(([key, value]) => {
+    newHeaders.set(key, value)
+  })
+  
+  return new Response(response.body, {
+    status: response.status,
+    statusText: response.statusText,
+    headers: newHeaders
+  })
+}