… __tests__/integration/r2-migration.test.ts
… dist/types/ranking.js
… dist/workers/api-gateway-r2.js
… docs/R2_ROLLBACK_PLAN.md