@YJSN180 2025-06-22
## ✅ CodeRabbitのレビューコメントに対応しました

以下の修正を行いました：

### 修正内容
1. **シェルスクリプトのエラーハンドリング強化** - `set -euo pipefail`を追加
2. **URLスキームの修正** - VERCEL_DEPLOYMENT_URLにhttps://を追加
3. **未使用importの削除** - zlibのimportを削除
4. **環境変数名の統一** - ドキュメントをR2_ACCESS_KEY_IDに統一
5. **バケット名の統一** - ドキュメントを`nico-ranking`に統一
6. **CIパイプラインの改善** - `build:worker`を`check:all`に追加

### 残課題
エラーハンドリングの部分的失敗対応については、現在の実装でも基本的な動作は保証されているため、将来の改善項目として残しています。

全てのCIチェックが成功していることを確認しました。